The data file HighLExtrapTemplate_lenspotentialCls.dat is loaded by relative path, meaning that fortran searches in the current working directory.
This means that running camb by /path-to-camb/camb (eg, from an external
script) fails unless this file is in the working directory of the
script.

This patch makes CAMB search the directory in which the binary is
located if opening by current working directory fails.

The code to do it is monstrous. Note however that it should work on
windows, due to searching for both / and \.