HMcode Dolag bug (#41)
In the previous version of CAMB the 1.5 power was applied to the 2015 version (imead=2) and the 1.0 power was applied to the 2016 version (imead=1), which was the wrong way round. This is now fixed.

This should only make a small difference at small scales for more extreme dark energy models. I wouldn't be able to quantify how small though.