bug in baseconfig.py
In the new version of the baseconfig.py file you appear to have introduced a bug. 

osp.join(BASEDIR, "..", "..", "fortran",
                     "HighLExtrapTemplate_lenspotentialCls.dat"))

in the previous versions of camb the above line was looking like this :

osp.join(BASEDIR, "..", "fortran",
                     "HighLExtrapTemplate_lenspotentialCls.dat"))

This make the python wrapper looking in the wrong folder for the HighLExtrapTemplate_lenspotentialCls.dat file.