Fix default argument list
getfullargspec parses the function argument list prepending `self` to the `args` list. So removing
first item is fine for `args` values but the `ðefault` argument values does not add a default value
for `self`. So there will be a shift by 1 when zipping both.

For cosmology parameters, for instance, the theta_H0_range is always skipped, same for sound speed
squared cs2.