Introduced active source correlator handling.
Created new python class that stores the eigenvector data calculated from active sources. This will need to be tied to a fortran class so that this data can be used for CAMB calculations.

Introduced the fortran ActiveSources module that takes in python data and makes it available for use in the CAMB calculations. Incorporated the sources into equations.f90

Added script to calculate cosmic string correlators and a test script to calculate the Cl-s from strings.